+++ CHANGELOG.md
@@ -5,6 +5,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Use Moka Cache for storing the sessions when no SQLite file is given.
+
 ## [0.1.0] - 2023-10-24
 
 Initial release of annis-web.